Add inject-dependencies transformer
